backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic-settings
python-dotenv
alembic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
